requirements-service.txt
pyyaml
tqdm
numpy<2.0
easydict
scikit-image
scikit-learn
opencv-python-headless
joblib
albumentations==0.5.2
imgaug==0.4.0
hydra-core==1.1.0
pytorch-lightning==1.2.9
tabulate
kornia==0.5.0
webdataset
packaging
omegaconf
flask
flask-cors
pillow
pandas
matplotlib
